refactor: tidies up tests
