+++ .eslintrc
@@ -20,7 +20,7 @@
     "import/no-named-as-default": 0,
     "arrow-body-style": ["warn", "as-needed"],
     "class-methods-use-this": 0,
-    "import/no-extraneous-dependencies": ["error", { "devDependencies": true }],
+    "import/no-extraneous-dependencies": 0,
     "no-param-reassign": ["error", { "props": false }]
   }
 }
