+++ package.json
@@ -4,13 +4,14 @@
   "private": true,
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
-    "test": "jest",
+    "test": "tape -r babel-register 'test/**/*.js' | tap-spec",
     "lint": "node_modules/eslint/bin/eslint.js 'src/js/actions/**/*.@(js|jsx)'",
     "test:watch": "nodemon -q -x 'npm test'"
   },
   "dependencies": {
     "@exponent/ex-navigation": "^2.1.5",
     "axios": "^0.15.2",
+    "babel-preset-es2015": "^6.18.0",
     "babel-preset-react-native-stage-0": "^1.0.1",
     "lodash": "^4.17.2",
     "moment": "^2.17.0",
@@ -38,6 +39,7 @@
   "devDependencies": {
     "babel-jest": "17.0.2",
     "babel-preset-react-native": "1.9.0",
+    "babel-register": "^6.18.0",
     "eslint": "^3.11.1",
     "eslint-config-airbnb": "^13.0.0",
     "eslint-plugin-import": "^2.2.0",
@@ -47,6 +49,7 @@
     "jest": "17.0.3",
     "nodemon": "^1.11.0",
     "react-test-renderer": "15.4.1",
-    "remote-redux-devtools": "^0.5.5"
+    "remote-redux-devtools": "^0.5.5",
+    "tap-spec": "^4.1.1"
   }
 }
