+++ test/unit/reducers/create-event/clear-create-event.js
@@ -1,5 +1,5 @@
 import test from 'tape';
-import reducer from '../../../../src/js/reducers/create-event.js';
+import reducer from '../../../../src/js/reducers/create.js';
 
 test('Reducer handles CLEAR_CREATE_EVENT as expected', (t) => {
