+++ test/unit/reducers/user.js
@@ -1,9 +1,5 @@
 import test from 'tape';
-import client from '../../../server/db/init.js';
-import reducer from '../../../src/js/reducers/user.js';
-import { server } from '../../utils/initDB.js';
-import { feedSocket } from '../../../src/js/socket.js';
-import { pub, sub } from '../../../server/init-socket.js';
+import reducer from '../../../src/js/reducers/user';
 
 
 test('Reducer handles GET_USER_REQUEST as expected', (t) => {
@@ -99,22 +95,4 @@ test('Reducer handles GET_USER_FAILURE as expected', (t) => {
 
     t.deepEqual(actual, expected, "GET_USER_FAILURE sets state correctly");
     t.end();
-});
-
-test('FINAL TEST TEARDOWN: test/unit/reducers/user.js', (t) => {
-
-    /* TEARDOWN: endpoints `new-event`
-    - decrement eventKeys
-    - delete event
-    - delete notification
-    - quit sockets and pub sub channels
-    */
-    feedSocket.disconnect();
-    pub.quit();
-    sub.quit();
-    client.del("vote:event:300|eventWhat:0");
-    client.del("vote:event:300|eventWhat:2");
-    client.del('notifications:12345678');
-    server.stop();
-    client.quit(t.end);
 });
