--- test/utils/shallowRenderer.js
@@ -1,11 +1,0 @@
-'use strict';
-
-import React from 'react';
-import { createRenderer } from 'react-addons-test-utils';
-
-export default function shallowRenderer (component) {
-
-    let shallowRenderer = createRenderer();
-    shallowRenderer.render(component);
-    return shallowRenderer.getRenderOutput();
-}
